refactor: move reaction function
